fixing unsigned integer problems, refs #1316
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@16505 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/output/MSMeanData.h
@@ -266,7 +266,7 @@ public:
                    const SUMOReal numLanes, const SUMOReal defaultTravelTime,
                    const int numVehicles = -1) const;
 
-        size_t getNumReady() const;
+        int getNumReady() const;
 
         void clearFirst();
